refactor(LottoWinningRank) : purchaseNumbersMatchWinningNumbersCounts 반환 타입 List<Long> 리팩토링
Long으로 purchaseNumbersMatchWinningNumbersCounts가 넘어오므로
List<Integer> -> List<Long>으로 리팩토링

+++ src/main/java/lotto/LottoWinningRank.java
@@ -4,19 +4,19 @@ import java.util.Arrays;
 import java.util.List;
 
 public enum LottoWinningRank {
-    NO_RANK(0, List.of(1, 2, 3), false, "0"),
-    FIRST(1, List.of(6), false, "2,000,000,000"),
-    SECOND(2, List.of(5), true, "30,000,000"),
-    THIRD(3, List.of(5), false, "1,500,000"),
-    FOURTH(4, List.of(4), false, "50,000"),
-    FIFTH(5, List.of(3), false, "5,000");
+    NO_RANK(0, List.of(1L, 2L, 3L), false, "0"),
+    FIRST(1, List.of(6L), false, "2,000,000,000"),
+    SECOND(2, List.of(5L), true, "30,000,000"),
+    THIRD(3, List.of(5L), false, "1,500,000"),
+    FOURTH(4, List.of(4L), false, "50,000"),
+    FIFTH(5, List.of(3L), false, "5,000");
 
     private final int rank;
-    private final List<Integer> purchaseNumbersMatchWinningNumbersCounts;
+    private final List<Long> purchaseNumbersMatchWinningNumbersCounts;
     private final boolean purchaseNumbersMatchBonusNumber;
     private final String winningMoney;
 
-    LottoWinningRank(int rank, List<Integer> purchaseNumbersMatchWinningNumbersCount,
+    LottoWinningRank(int rank, List<Long> purchaseNumbersMatchWinningNumbersCount,
                      boolean purchaseNumbersMatchBonusNumber, String winningMoney) {
         this.rank = rank;
         this.purchaseNumbersMatchWinningNumbersCounts = purchaseNumbersMatchWinningNumbersCount;
@@ -28,7 +28,7 @@ public enum LottoWinningRank {
         return rank;
     }
 
-    public List<Integer> getPurchaseNumbersMatchWinningNumbersCounts() {
+    public List<Long> getPurchaseNumbersMatchWinningNumbersCounts() {
         return purchaseNumbersMatchWinningNumbersCounts;
     }
